Beta Version 0.9
- Added order confirmation that will update the db, marking today's order as complete and inform a future user to check back tomorrow.

- Edited sales button on orders submenu to lead to a new Activity, ReportSalesActivity that will enable the user to get sales reports.
TODO: Ensure the user can't pass a wrong format and create functionality.

From this point on, the application can perform most of the intended functionality. Not ever tab does something. For example by creating the sales report, one can assume a similar functionality on the rest of the reporting tab and therefore, for time's sake, the rest will be skipped. Since the dissertation project is a conceptual project, the important part is to specify ideas rather than a complete product ready to ship.